+++ src/state/include/state.h
@@ -443,7 +443,7 @@ public:
 	 * @param[in] delta_time The difference in time in ms between
 	 *                       the previous and current Update calls.
 	 */
-	void Update(int64_t delta_time);
+	void Update(float delta_time);
 	/**
 	 * Merges this, the main state, with a player's copy of the state
 	 *
